[BUGFIX] Invalid behavior of search for integer in Backend search
In fact current implementation does not replace the LIKE
operator with an SQL equal statement for numeric fields
but whenever a query that can be interpreted as integer
is encountered, regardless of the underlying column type.

Example: If you have pages "102 W", "102 E", you used
to be able to search for "102" and get them, but now
the query is casted to an integer and the page title
is search for an equality on "102" resulting in
an empty result list.

The underlying column type should be used
instead (code to cache this information is already
available within DBAL and Helmut started months
ago to backport it to Core itself: #28654).

Thanks to Clio Online for sponsoring this fix!

Change-Id: I527b2eb742f663985866e717fcf572f2a7e33574
Resolves: #33700
Releases: 6.0, 4.7, 4.6