[BUGFIX] Preview images not rendered in Web > Page view on Windows systems
In Web > Page view the preview images are not rendered on Windows systems.
This is due to failing functions in class
TYPO3\CMS\Core\Utility\PathUtility. Basically it's an issue with forward
and backward slashes. In the class the constance DIRECTORY_SEPARATOR is
used which is on Windows systems a backward slash. Normally paths in TYPO3
are converted to use forward slashes and here is the existing problem.

This patch removes the constance in some functions (which are nessessary
for preview images) and makes sure all pathes are converted to unix style.
Than processing can use forward shlashes and carry on.

Change-Id: If7b443e5180e26ceb089a5b47c1631c4c7eb21c1
Fixes: #41204
Releases: 6.0, 4.7
Reviewed-on: https://review.typo3.org/17622
Reviewed-by: Mattias Nilsson
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn