[BUGFIX] Deleted processed files are not rewritten
Files which were already processed for some context (e.g. preview of
images) are not rewritten when you delete the processing folder. This
happens because only database records are evaluated but the file itself is
not checked. This patch adds a check for the existing file for processed
flag.

Change-Id: I16713ba43e730dcb012a20072c9f8261576cd720
Fixes: #41223
Releases: 6.0