[BUGFIX] Deprecate t3lib_install
t3lib_install contained two logical parts in the past: A group of methods
to handle localconf.php and another group to parse ext_tables.sql files
and friends. The latter was already extracted to t3lib_install_sql in the
past. The localconf handling is obsolete with the new
LocalConfiguration.php and its new API method.

The patch handles:
 * Deprecate t3lib_install and change the code in a way that does not
   harm if still called.
 * Move some methods and properties to TYPO3\CMS\Install\Installer, make
   the class standalone and do not extend t3lib_install anymore.

Fixes: #41245
Releases: 6.0

Change-Id: Iafd856ecdc9e158fdc1b31d1f7be97064234ba6c