[BUGFIX] BEfunc getProcessedValue doesn't show record title for "group"
In the list view relation fields using the type group (internal_type=db)
doesn't display the related value as expected. Instead, it shows
the UID of the relation. As these calls go through getProcessedValue,
the issue resolves the records by displaying the record titles correctly
in the TYPO3 backend.

Change-Id: I7d9cc8ec7f02e4a02518fcd14c81cc71f69b8ea7
Resolves: #21410
Releases: 6.2, 6.1, 6.0
Reviewed-on: https://review.typo3.org/20521
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters