[BUGFIX] BEfunc getProcessedValue doesn't show record title for "group"
In the list view relation fields using the type group (internal_type=db)
doesn't display the related value as expected. Instead, it shows
the UID of the relation. As these calls go through getProcessedValue,
the issue resolves the records by displaying the record titles correctly
in the TYPO3 backend.

Change-Id: Idc2eabbafc61726b122ced459346a87f3b9c6543
Resolves: #21410
Releases: 6.0