[TASK] Improve show_item for FAL files
The show_item.php shows some info about the selected file.
the property size is rendered twice, once in the
header section with a size (eg. 223kb) and below with
the size as byte. remote the last one.

A call to trim() can be removed as trimExplode() is used before

Change-Id: I33b0d9270bc6f81d490348c050c2b9e30c4a102c
Resolves: #41291
Releases: 6.0