[BUG] Shared defaults for CE wizard using TCAdefaults
The page TSconfig settings for the New Content Element wizard have
a higher priority than the user TSconfig TCAdefaults. Settings which
are not different for each CE type should be set in TCAdefaults. If
they are set in the tt_content_defVals the values will be different
if the editor selected the right CE type directly than if the editor
switched CE type.

Change-Id: I8b84b3059b17f88a3d674b6faae889844cae9c67
Fixes: #41434
Release: 6.0, 4.7, 4.6, 4.5