[BUGFIX] Root-level folders cannot be displayed
Due to a recent fix in the folder resolving mechanism, the root-level
folders of a storage were not displayed anymore in the file list module.

This commit re-enables displaying rootlevel folders, at the expense of
removing the hierarchy resolving (i.e., we now always choose the
root-level folder of a storage instead of the next higher level. This
concept was broken anyways, as folders do not need to be in a hierarchy.
Thus we cannot use such a simple resolving mechanism, but have to
implement that per driver instead.

There is currently no permission check in the file list controller, so
both the selected folder and the (fallback) root-level folder might be
displayed though the user has no permission to do so.

Change-Id: I71580031f0cc14b2cfac76b80b3641dd05a3e052
Resolves: #41569
Releases: 6.0
Reviewed-on: http://review.typo3.org/15321
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel