[TASK] Make Core\Configuration\ConfigurationManager non static
To have the possibility to remove dependencies from code that
calls the configuration manager, it has to be made non static.

Together with this change, all further configuration file logic
is moved to that class, to not expose internals in other
places. For that some methods needed to be made public, but
have been marked private, thus are not public API.

Further cleanup and simplification, especially not calling
makeInstance several times, will go in a separate change.

Resolves: #41052
Releases: 6.0

Change-Id: I8ddfa4a957d0b4955ed422a1deeebaf84340c2c1