[BUGFIX] Consider localizeReferencesAtParentLocalization for group/db
Currently the TCA config property
localizeReferencesAtParentLocalization for database relations is
only supported for select types with an accordant foreign_table
defined. However, the whole procedure is valid but not checked
for group/db fields as well.

This change considers now all allowed tables and checks whether
at least on of them is localizable. If one table cannot be
localized just the current value is taken as is.

Change-Id: Iff09f261040f245d503c8273871b2a37c0723a80
Fixes: #41713
Releases: 6.0