[TASK] Improve show_item for FAL images
Reverts the existing change, and makes it more flexible by
modifying the showRecordFieldList instead of hardcoding the
exclusion of the "size" property.
The show_item.php shows some info about the selected file.
the property size is rendered twice, once in the header section
with a size (eg. 223kb) and below with the size as byte.
remote the last one.

Change-Id: I1aa26bb8162195cd20f8756d944f862af18b3bcd
Resolves: #41291
Releases: 6.0