[BUGFIX] Folder: File list filtering does not work
To filter a file list for certain file extensions, the
FileExtensionFilter class should be used. However, the element browser
currently tries to directly call getFiles() with a file extension list.

This commit adds filter handling methods to Folder and changes the EBs
behaviour to use them. The public interface of Folder::getFiles() and
Folder::getSubfolders() is extended, but in a backwards-compatible way,
so this is no breaking change.

Change-Id: Ib45d5f0a99c038841193ed568d7ebf8289f629d3
Resolves: #41725
Releases: 6.0