[BUGFIX] MEDIA element ignores audio fallback
In TYPO3 4.7 the field name containing the reference to an audio
file was changed in the MEDIA element, but the renderers were not
adapted (except the flowplayer). Other renderers must also consider
"audioFallback".

Additionally for the QT renderer, the type must be evaluated
a bit earlier, so that it can be used for audio fallback calculation.

Change-Id: Ie00d5fafc780e53fb44c02f0ff2ade45359f1bcc
Resolves: #39416
Releases: 4.7
Reviewed-on: http://review.typo3.org/15412
Reviewed-by: Lorenz Ulrich
Tested-by: Lorenz Ulrich
Reviewed-by: Steffen Ritter
Tested-by: Steffen Ritter