[BUGFIX] FlexFormTools does not use UTF-8 by default
When using FlexFormTools, and no $LANG->charSet is
set, then the XML header is written with an empty
character set. This happens if you use this function
in your own extension and not in the backend e.g.,
but also when instantiating the storages
(this is where I noticed it).

Thus, we introduce a fallback to UTF-8. This is
the same way, GeneralUtility (t3lib_div) -
array2xml_cs does it.

Change-Id: I2a83b155d164ce5c4076b06bfbc219d7dd09fc38
Releases: 6.0
Resolves: #41793