[BUGFIX] FAL: Default storage is created multiple times
The "auto-creation" of the first file storage should only be
triggered once. Currently this happens whenever there is no
record for a user (thus a loop).

By checking the DB table for any record, which is done once
before the actual DB query, the auto-creation is really done
just a single time - when sys_file_storage is empty.

Change-Id: I6cd84bb68af15caffa3185f0a8541dcb30eecdaf
Releases: 6.0
Resolves: #39785