[BUGFIX] RTE: Link to disabled page doesn't show in FE, link icon does
Problem: When linking from a piece of content in the RTE to a disabled
page in the tree, the link isn't shown in the final frontend rendering.
However, the small arrow icon that you can have in front of the link to
denote its type, does get shown.
Solution: Remove the icon if no link is generated

Change-Id: If2f5af7ef81578571ed0a1255a53d9373f14bfed
Resolves: #36087
Releases: 4.5, 4.6, 4.7, 6.0