[BUGFIX] RTE: Link to disabled page doesn't show in FE, link icon does
Problem: When linking from a piece of content in the RTE to a disabled
page in the tree, the link isn't shown in the final frontend rendering.
However, the small arrow icon that you can have in front of the link to
denote its type, does get shown.
Solution: Remove the icon if no link is generated

Change-Id: I5199d90aaf4f2b3103ff1f8680376b82173e9723
Resolves: #36087
Releases: 4.5, 4.6, 4.7, 6.0