[BUGFIX] RTE: Link to disabled page doesn't show in FE, link icon does
Problem: When linking from a piece of content in the RTE to a disabled
page in the tree, the link isn't shown in the final frontend rendering.
However, the small arrow icon that you can have in front of the link to
denote its type, does get shown.
Solution: Remove the icon if no link is generated

Change-Id: I32a1eecf3bba474c8b7617bae28362abc561e8b2
Resolves: #36087
Releases: 4.5, 4.6, 4.7, 6.0
Reviewed-on: http://review.typo3.org/15739
Reviewed-by: Mattias Nilsson
Tested-by: Mattias Nilsson
Reviewed-by: Stanislas Rolland
Tested-by: Stanislas Rolland