[BUGFIX] Set class alias also for new class names
As not all usage of class names trigger the autoloader,
usage of old class names in type hints for example doesn't.

If an object with a namespaced class name is given to
a method with an old class name, PHP dies with a fatal error.

To ensure full compatibility we need to set the
class alias also when a namespaced class is required
in the autoloader.

Resolves: #42058
Releases: 6.0

Change-Id: Ia4babeeae07fb778732c5846c5cf8c70115037f0