[BUGFIX] Storage uses wrong identifier for processing folder
All absolute identifiers within a storage have to begin with a "/". The
Storage does not respect this for the processing folder, leading to
wrong identifiers being generated for images inside this folder.

To account for this change, the sys_file_processedfile table should be
cleared. All identifiers will be regenerated then.

Change-Id: I19293ea766f54830264555f63094471f73a61b28
Resolves: #42180
Releases: 6.0