[TASK] Remove excludefields at sys_file_reference table
The TCA for sys_file_reference table configures fields like
hidden, sys_language_uid and alike as "exclude => 1".

This causes the translation of tt_content elements with
images to fail for an regular editor if these fields are
not configured to be allowed explicitly.

These fields are always hidden in TCEforms, but always
required for IRRE and translation to work.
Therefore remove the exclude-field configuration.

Change-Id: I1783b7406b62c1142acb991308e3698cf5af68f1
Resolves: #42224
Releases: 6.0