[BUGFIX] Images from non-local storages are not displayed
The content object renderer adds absRefPrefix to all generated image
URLs, leading to incorrect URLs for storages not relative to the
siteroot. Additionally, the URLs generated by the driver are also
escaped in ContentObjectRenderer::cImage(), leading to broken absolute
URLs.

The prefixing with absRefPrefix is now done in LocalDriver and only for
relative filemounts.

This patch also moves the escaping to the LocalDriver; it should be
added as a requirement for other drivers.

Change-Id: I40c0a7ea55ff3a1155e7a05b1fc37da7d0cf44e4
Resolves: #42223
Releases: 6.0