[BUGFIX] Hidden files/folders are always filtered
The local driver has a hardcoded filter for dot-files, i.e. files whose
name starts with a dot. However, we (or the user) might need these files
in some situations. Therefore, the filter should be added by default,
but should be removable, which is the case if we use the extensible
filter mechanisms provided by FAL. This is done with this commit.

Resolves: #42230
Releases: 6.0
Change-Id: I69ff38d35f75e2a7f4bbb7574caf6a261638512d