[BUGFIX] FAL: Misleading default filter array name
The key for the default filters array in the default configuration of
TYPO3 is callbackFilterMethods. This is misleading, as one might see
this as a list of *all* available filters. Therefore, the key is renamed
to defaultFilterCallbacks.

The name change does not break any public API as the array key is only
used in one place: the ResourceStorage.

Resolves: #42239
Releases: 6.0
Change-Id: I48282f8e49ab786aaf8821badfb411a653e9a3fc
Reviewed-on: http://review.typo3.org/15856
Reviewed-by: Steffen Ritter
Tested-by: Steffen Ritter