[TASK] Database creating fails due to not allowed characters
Due to compatibility the CREATE DATABASE statement is used without quotes.
At least for MySQL databases there are limited characters allowed to use
in unquoted database names (0-9,a-z,A-Z$_). At the moment the database
name is parsed with enabled hyphen but this is obviously not allowed. The
parsing has to be changed to meet the MySQL limitation as well.

Change-Id: I24dcf99bb9baed6a4794183de0eb4f99a0948b7c
Resolves: #41151
Releases: 4.5, 4.6, 4.7, 6.0