[TASK] Add possibility for a custom TCA TreeProvider
Currently the tca tree uses its own DataProvider
which makes sense but this is not extendable.
Adding 3 additional lines would make it possible
to use a custom provider.

Needed e.g. in EXT:news to be able to render a tca tree
with some BE user/-group restrictions ...

As this is really a small change I would love to see
that also in lower versions

Change-Id: Idc9bee2ede810e930c17a31ed08688b5654759ff
Resolves: #42447
Releases: 6.0, 4.7, 4.6, 4.5