[BUGFIX] Fix shown references of files
Currently the references of files are only fetched from
the sys_file_reference table. This does not take all direct
references and soft references into account.

Besides that, the count shown in the list is taken from the
count of tooltip hints instead of real references.

Fix both issues with this change.

Fixes: #42453
Releases: 6.1, 6.0

Change-Id: I82995896b5d62ee8df7baa175a752a6a25ce9c24
Reviewed-on: https://review.typo3.org/15986
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel