[BUGFIX] Fix handling of additionalHeaderData in USER_INT
While introducing USER_INT handling for the pageRenderer
a bug has been introduced which breaks traditional usage of
additionalHeaderData and additionalFooterData.

This is the case because the replacement of the headerData is done
first, when the to be replaced strings are not present
in the HTML content.
They are stored in the pageRenderer context.

Because of that, we need to replace the pageRenderer marker first
and after that the marker for additionalHeaderData
and additionalFooterData

Fixes: #42490
Releases: 6.0

Change-Id: I91ef658b5b5c8ca5c6cf7e25752e59a33152c2ac
Reviewed-on: http://review.typo3.org/16013
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel