[BUGFIX] Restricted elements should show edit options
If a content element is not allowed to be edited, e.g.
because the ctype is not allowed to the current backend user,
the edit options should not be shown. Also the content
element should not be linked - the form is not displayed
anyway but an ugly error is shown.

Change-Id: I94beefdf6fc6a910201ab8fba4dbda0c5d174359
Fixes: #42501
Releases: 6.0