[BUGFIX] Update LocalConfiguration fails (Install Tool: Upgrade Wizard)
Due to changes in function
TYPO3\CMS\Core\Configuration\ConfigurationManager::writeLocalConfiguration
an error is thrown when the configuration file isn't accessable. Of course
when updating the system from a lower version it is not available. So the
installation process has to create the file on its own to be able to write
the settings.

Change-Id: I89553390ebd9f1f576311535e02ec73da6618cd6
Fixes: #42271
Releases: 6.0