[BUGFIX] Translated non-published page in workspace breaks live workspace
Translation of pages are stored in pages_language_overlay and
can be managed in workspaces as well. If a new translation is
created in the workspace only, the page module is broken in
the live workspace.

The reason for that is, that workspaces placeholders need to
be ignored if not working on a real workspace.

Change-Id: Ied52a985c3cb1c0796f616679414799d956024d8
Releases: 6.0, 4.7, 4.6, 4.5
Fixes: #42281