[!!!][TASK] Rename ExtensionManager class
The class ExtensionManager is actually a utility
to manage extensions. Rename it to better reflect
it's purpose.

This is the final commit, which removes the old
namespaced class name from the alias map and
raises the submodule pointers.

If you used ExtensionManager calls in your TYPO3 6.0
extensions, you have to adapt your code and use the
new class name.

Resolves: #42661
Releases: 6.0

Change-Id: I371a64b276a93f25fabf6728bae8906fbc0e2c42
Reviewed-on: http://review.typo3.org/16176
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel