[BUGFIX] deleteRecord_flexFormCallBack deletes file_references
Files that are referenced in a flexform
<type>group</type><internal_type>file_reference</internal_type>
will be deleted by deleteRecord_flexFormCallBack() in
t3lib_TCEmain, independent of the path or usage by other records.

For normal record fields (e.g. tt_content.images), only file
with the internal_type "file" are deleted. Files of
internal_type "file_reference" are respected and will not been
deleted. There is an appropriate condition in
t3lib_TCEmain::extFileFunctions().

But for flexform fields, this differentiation in handling
between both internal_types is missing.

Therefore using lowlevel_cleaner or the delete option in the
recycler will delete files that may be still in use. The
flexFormCallBack must be changed to delete only files of
internal_type "file".

Change-Id: I8519705f2155bf0d49ae10316c6622f670189e28
Releases: 6.0, 4.7, 4.6, 4.5
Fixes: #40052