[FEATURE] Enable IRRE fields in FlexForms
This patch enables IRRE fields in flexform. They can be used
identically to the TCA.

FlexForm segments are separated by a new divider "---" to
easily have the possiblity to recognize and parse them when
handline the IRRE object identifiers.

Besides that a new remapping level for FlexForms has been
introduced to the DataHandler to update accordant FlexForm
references after all child records have been written.

Unit tests have been extended to check the new additions to
the accordant object identifiers and form names. This is
important to any processing of IRRE in the form view.

Change-Id: Icadb89c2e496a5f1ad7de298ebee06d144475a11
Resolves: #18957
Releases: 6.0