[TASK] Adjust phpdoc for classes (files) of backports
Since it's neither "FLOW3" nor "v5" anymore those
headers need to be adjusted.

Before:
This class is a backport of the corresponding class of FLOW3.
All credits go to the v5 team.
Changed into:
This class is a backport of the corresponding class of TYPO3 Flow.
All credits go to the TYPO3 Flow team.

Change-Id: If5aaae848334c39a873b41ffc614942cedbbf4f8
Resolves: #42713
Releases: 6.0
Reviewed-on: http://review.typo3.org/16242
Reviewed-by: Stefan Neufeind
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring