[BUGFIX] Fix namespaced \Countable
Usage of namespaced Countable call breaks the compatibility for
PHP 5.2.

Besides that unit-test have been modified to fit the interface
of Extbase 1.3 which does not support validate() and hasErrors()
for validators. The SplObjectStorage (which was introduced with
PHP 5.3) was replaced by ArrayObject since an implementation
against the Countable interface is relevant here.

Change-Id: I26a861a20e67a07118dc4d52fcc1cfdba8aea50c
Fixes: #42811
Releases: 1.3
Reviewed-on: http://review.typo3.org/16344
Reviewed-by: Stefan Neufeind
Reviewed-by: Steffen Ritter
Tested-by: Steffen Ritter
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader