[BUGFIX] Fix problem with non accessible pages and PageNotFound handler.
This bug exists since 3.8.0 or earlier.
The var pageNotFound is set, if the called page has access restrictions.
Later on we try to find another page uptree but we do not reset the
pageNotFound var if we found an accessible page.

The PageNotFound handler reacts on this var and redirects to the 404
page instead of presenting the accessible page we found later on.

Resolves: #16472
Releases: 4.5

Change-Id: I021410fa7a6a9507b333c0c6a01dc49a11b8a329