[BUGFIX] Content Elements "uploads" ignores translations
The TypoScript Object "FILES" has been changed to be aware
of translations in the meantime. Anyhow the Content Element
uploads does not benefit from that, as it hardcodes the uid
of the related element and therefore the automatic logic
recognizing translations does not work.

Remove the affected line of TypoScript.

Change-Id: Ia70272a2d646c2d29eca311c44dc6a0706d1778c
Releases: 6.0
Fixes: #40607
Reviewed-on: http://review.typo3.org/16363
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader