[BUGFIX] Fix userfunc paths in EXT:saltedpasswords ext_conf_template.txt
Paths to the class files and class names for the user functions used
in the extension configuration of EXT:saltedpasswords are wrong, since
they were moved and namespaced.

Thus opening the extension configuration in extension manager fails
with a fatal error, because the required class files can not be loaded.

Load the right classes.

Fixes: #42829
Releases: 6.0

Change-Id: I9e9dd6f3f8ceacf0196971db91f108d7dc963bcd
Reviewed-on: http://review.typo3.org/16368
Reviewed-by: Philipp Gampe
Tested-by: Philipp Gampe
Reviewed-by: Mattias Nilsson
Reviewed-by: Tilo Baller
Tested-by: Tilo Baller
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn