[BUGFIX] Fix userfunc pathes in EXT:saltedpasswords ext_conf_template.txt
Pathes to the class files and class names for the user functions used
in the extension configuration of EXT:saltedpasswords were wrong, since
they were moved and namespaced.

Thus opening the extension configuration in extension manager failed
with a fatal error, because the required class files could not be loaded.

Fixes: #42829
Releases: 6.0

Change-Id: I9e9dd6f3f8ceacf0196971db91f108d7dc963bcd