[BUGFIX] Fix missing option labels in EXT:saltedpasswords ext config
The labels for the options of the settings "Hashing method for the
backend|frontend" are missing because the class names of the
salt methods, which changed in previous commits, are part of the
language key for the labels.

Change the language key generation to use the class name without
the prepended namespace and add the corresponding language key in the
actual language file.

Fixes: #42832
Releases: 6.0

Change-Id: I1ae96bd90bf153f0379159e1695701ecf5f9f004