[BUGFIX] Fix update translation handling
The current translation handling does not respect
subfolders in the zip file during extraction.
This leads to partly missing translations as the
translation files were not in the correct folder
structure.

Create the folder structure during zip file extraction
and improve the path calculation so that path traversal
is not possible any more.

Fixes: #42840
Releases: 6.0

Change-Id: I9e03622879f38d294b2487c63e9ced0ba40d8142