[BUGFIX] Fix charset inconsistency in Fluid tests
The encoding viewHelper calls resolveDefaultCharset
which in turn tries to get the charset from the system.
This is incomplete mocking and is resolved by mocking
this method. Doing so shows up another inconsistency.

Some tests assume ISO-8859-1, some UTF-8 to be the
default charset. Fix the tests to always assume UTF-8
to be the default encoding.

In addition remove the unnecessary checks in
resolveDefaultEncoding as the system default is UTF-8

Resolves: #43075
Releases: 6.0

Change-Id: Id311e5e3621c1c105a2cd182881265b92cedfebf
Reviewed-on: http://review.typo3.org/16557
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel