[BUGFIX] Catchable fatal error when using the swap button
An error message is coming up if you push the swap button (double arrow).
The swap is realy done but you have to address the record once
again in the list or page view, because a error message
is still shown in the working frame on the right.

To reproduce this error select a page with multiple versions in a page or
list module in live workspace. You find the version module in the top of
right frame. Press the 'Version Management' button. Now you get the swap
button in front of each version.

Fixes: #42948
Releases: 6.1, 6.0, 4.7, 4.6, 4.5

Change-Id: I45b3e691932782d6e4085c989b1fe0809b5a34ab
Reviewed-on: https://review.typo3.org/18013
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn