[BUGFIX] AbstractFormFieldViewHelper always converts entities
AbstractFormFieldViewHelper::getValue() currently always tries
to convert objects to identities by calling
PersistenceManager::getIdentifierByObject().
This makes it impossible for custom ViewHelpers to access the
unmodified value that was previously submitted (or bound to the
form field view helper).

This change introduces an argument $convertObjects that allows
experienced developers to disable this behavior in custom
view helpers.

Fixes: #34091
Fluid Issue: #33569
Releases: 1.3, 1.4, 4.7, 6.0

Change-Id: I72ffbfb6238cf0d12301687750f0314b97f2faad
Reviewed-on: http://review.typo3.org/9098
Reviewed-by: Marc Bastian Heinrichs
Tested-by: Marc Bastian Heinrichs
Reviewed-by: Ingo Pfennigstorf
Tested-by: Ingo Pfennigstorf
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring