[BUGFIX] Rootline relation overlay does not work for translated pages
Overlays for the relation fields of a page in the rootline, introduced
with the rootline class refactoring, do not work for translated pages.
The problem is that the overlay of data is done too early:
RootlineUtility::enrichWithRelationFields() already kicks in long before
the sys_language_uid is determined.

This patch does additional overlays in the TypoScriptFrontendController
when the rootline is being updated. In addition, the method in question
is moved from the RootlineUtility to the PageRepository.

Change-Id: If32830b384b08e189d155532aee7281b040b7dd5
Resolves: #42029
Releases: 6.0