[TASK] Prepare AbstractRepository of FAL for FLOW sync
The extbase team wants to be in sync with TYPO3 Flow
again which leads to the change of the Repository
Interface that defines three new methods one has to
implement to fulfill the interface contract. As the
AbstractRepsoitory of FAL directly uses the interface
instead of the Repository class it has to implement
these methods itself.

Releases: 6.1
Fixes: #43092
Change-Id: I14a1c8f3111c55f4a7b77d6c82669073b5a4a150