[TASK] Remove extList from LocalConfiguration
extList was kept next to the new extListArray in LocalConfiguration.
Both hold the current list of loaded extensions. This leads to
confusion which variable is the one that takes precedence.
The patch removes extList from the upgrade wizard in the install tool
and only writes extListArray. This way, only one key is kept in
LocalConfiguration.php. For compatibility reasons, they array entry
'extList' is added to $GLOBALS['TYPO3_CONF_VARS'] at runtime during
bootstrap.

Change-Id: I6102b2b4da9761c86189ae3f73ad3a423715e037
Resolves: #43094
Releases: 6.0