[TASK] Integrate possibility to call inaccessible methods in BaseTestCase
The BaseTestCase already has a possibility to create full blown
accessible object proxies. However if one does not want to have
the class stack cluttered with temporary existing classes, there
shall be a possibility to directly invoke protected/private
methods of a fixture object.

The new method callInaccessibleMethod method is taken and
rewritten from ModuleControllerTest.

Example to use it in unit tests:
$this->fixture = new \TYPO3\CMS\Core\SomeClassName();
$result = $this->callInaccessibleMethod(
  $this->fixture,
  'privateMethod',
  $first, $second, $third, ...
);

Change-Id: I2c90c136c350129b28809d8b8b88286149894a10
Resolves: #43145
Releases: 6.0
Reviewed-on: http://review.typo3.org/16598
Reviewed-by: Philipp Gampe
Tested-by: Philipp Gampe
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader