[BUGFIX] Fallback of FAL properties does not work
The fields title, description and alternative are available
in file references and files. If those in file references
are not defined, those of files shall be used.

Besides that, some fields are not configured yet to be shown
as placeholder element in the backend.

The flaws are fixed by adding accordant TCA configuration and
adding the fallback behaviour for the FileReference object.

Change-Id: I28fa94faeb88958ce6c6cb942ba9a88bd555acdf
Fixes: #43173
Releases: 6.0