[BUGFIX] Deactivated textarea elements are not visualized
Deactivated textarea elements, fields that have a NULL value
stored in the database, are not visualized correctly when
loading the backend editing form. The status is currently set
using JavaScript and triggered by TBE_EDITOR.fieldSet().
However, textareas don't have a human readable clone and thus
fieldSet() is not called and thus no status is set.

This issue is fixed by directly rendering the status in the
FormEngine.

Change-Id: I2dded816853f452b25f6edf6e7d7f6dca1387739
Fixes: #43175
Releases: 6.0
Reviewed-on: http://review.typo3.org/16636
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader