[TASK] Allow Folder and Storage to recursively retrieve Files from Driver
The drivers in FAL are capable of retrieving files recursively.
This might be sinful in some cases and custom usages. Anyhow,
the parameter $recursive is not passed up until Storage and
Folder objects. With that users are forced to work with the
driver directly.

As it is highly discouraged to work directly on the driver,
just pass up the parameter within the abstraction layers.

Change-Id: Ibe8f9aa9e906617a42a9a7d5edba43f24f613587
Releases: 6.0
Fixes: #43249
Reviewed-on: http://review.typo3.org/16697
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel