[TASK] Change Xclass register from ux_ prefix to Xclass suffix
With 6.0 a new Xclass registration scheme was put in place removing
the global TYPO3_CONF_VARS['XCLASS'] registration and using the
autoloader instead.
Currently, an Xclass is registered by prefixing the class that has
to be Xclass'ed with 'ux_'. An example is in dbal
'ux_TYPO3\CMS\Core\Database\DatabaseConnection'.
This looks ugly and doesn't really make much sense with namespaces.
Therefore, a late change is introduced, changing the 'ux_' prefix to
a 'Xclass' suffix.

Example for a Xclass registration of a namespcaed class,
makeInstance will return an instance of the Xclass class, if asked
for an instance of the orginial path:
'Original\Classname' => 'Other\ClassnameXclass'

Non name spaced expample:
'tx_FooBar_Classname' => 'tx_FooBar_ClassnameXclass'

The patch adjusts the ClassLoader and makeInstance. Additionally,
two unit tests for the ClassLoader are removed that duplicated
other tests and did not make sense.

Resolves: #43269
Releases: 6.0

Change-Id: I04198b178aa41f02d7a5b7ee990ff24f483990f5