[TASK] Reintroduce Xclass registry
With 6.0 all Xclass inclusion in the classes have been
removed and registration of Xclasses has been moved to
the autoloader. This added additional complexity and
another concern to the class loader.

The Xclass registry will now be re-introduced but giving
it another drive, by only storing the relation between the
original class name and the target class name.

The path of the target class can be resolved automatically
by the class loader if naming conventions are met or the
target class name is put into the autoloading registry.

An Xclass can now be registered like that:

$GLOBALS['TYPO3_CONF_VARS'][TYPO3_MODE]['XCLASS']['originalClassName'] = 'targetClassName';

The patch adjusts the ClassLoader and makeInstance.
The unit tests for the ClassLoader are removed since
it is not responsible for resolving Xclasses any more.

Instead unit tests for makeInstance are added to test
the new behavior.

Resolves: #43269
Releases: 6.0

Change-Id: I04198b178aa41f02d7a5b7ee990ff24f483990f5