[TASK] Protect bootstrap methods
With 6.0 the bootstrap related core code was split into small methods
and transferred to a group of encapsulating classes.
While this is an important step to get a flexible and maintainable
bootstrap in the end, this process is not completed and still misses
for example a real concept for scopes.
The patch groups methods used in all scopes in bootstrap wrapper
methods and makes the business methods protected. The whole API and
all affected classes are marked as "internal", together with a
warning that this API will change in the future and shouldn't be
used by 3rd party code that is not under core control.
This gives freedom for future development of this core code without
taking care of backwards compatibility. With previous versions there
was no API at all, so this is not a feature loss from an extension
point of view.

Change-Id: Ib841a19c766691fdeb7dab07280a005d6b67a938
Resolves: #43285
Releases: 6.0