[BUGFIX] Disable some not finished functions in EM
The Extension Manager uses jQuery DataTables, but the
current implementation is not streamlined and causes
more trouble than it helps. Same goes for the tooltips
which are not properly styled.

As discussed with the UI/UX team we disable this
functionality and properly implement it in later
versions as bugfixes.

Fixes: #43319
Releases: 6.0

Change-Id: I432a15bce3c3933efa33f65304b01e2720b051a5