[BUGFIX] Wrong column titles in web_list if using backend_layout
If adding new content elements in Web>List with the CE wizard,
you get the old column titles even if you have a complete
backend_layout working.

But the Backend Layout is already correct. Only the titles are
wrong! Column titles are already in $tcaItems a few lines before
the lines in the patch.

Using this values the backend_layout is correct in Web>List
CE Wizard. Also for Moving Items Bug #24992.

Resolves: #42457
Releases: 4.7, 4.6, 4.5

Change-Id: Iac309b1aada4c5d74f30f77dfb1fe06ebcda3157