[BUGFIX] Fix column layout for New Content Element wizard
If adding new content elements in Web>List with the CE wizard,
you get the old column layout & titles even if you have a complete
backend_layout working.

But for Moving Elements the Backend Layout is already correct.
Column titles are already in $tcaItems a few lines before
the lines in the patch for File PagePositionMap.php. Using this
values the backend_layout is correct in Web>List CE Wizard.

Resolves: #42457
Releases: master, 6.2
Change-Id: I5f57aeac493ec5db02aaa92bd9eab97390cdb967
Reviewed-on: http://review.typo3.org/33644
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>