[BUGFIX] Fix column layout for New Content Element wizard
If adding new content elements in Web>List with the CE wizard,
you get the old column layout & titles even if you have a complete
backend_layout working.

But for Moving Elements the Backend Layout is already loaded -
even if title are wrong. The correct column titles are already
in $tcaItems a few lines before the lines in the patch for File
class.t3lib_positionmap.php. Using this values the backend_layout
is correct in Web>List CE Wizard.

Resolves: #42457
Releases: 4.7

Change-Id: I929987e29aee9d878623985e48f6657e656d1c1a