[BUGFIX] ContentObjectRenderer instantiates old class names
The content object renderer still instantiates content objects with
the tslib_content_ prefix. The patch changes this to the namespaced
class. Additionally, all content objects are now tested for successful
instantiation with additional unit tests. While creating those tests
it became obvious that the ShockwaveFlashObject content object was
named incorrectly, so this is adapted along the way.

Change-Id: Ie12474df270b29e8e2d8d8894fa94d24747d94d8
Fixes: #43456
Releases: 6.1, 6.0