[BUGFIX] Image viewhelper clears $GLOBALS['TSFE'] in backend context
The image viewhelper replaces the $GLOBALS['TSFE'] array to
simulateFrontendEnvironment if it is called in backend
context. After successful get of the image, it restores the
TSFE.

But in case, that the requested image has not been found,
it throws an Tx_Fluid_Core_ViewHelper_Exception, which means
that the TSFE get not restored. This causes errors when
other code tries to access TSFE after the image viewhelper
has been executed.

Resolves: #43446
Release: 6.0

Change-Id: I7563aef8481cacb602506eed557b31f6dfb3f315
Reviewed-on: https://review.typo3.org/16942
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring