[BUGFIX] Fix rootline translation
The fix in #42029 introduced a regression
as TSFE->tmpl->rootLine holds a reversed
rootline array, while in TSFE->rootLine
the order is "normal"

Additionally the call to getPageOverlay was wrong
because the result of the overlay was never used.

Lastly use the correct uid for looking up the
relation fields.

This change also cleans up the unit tests for
RootlineUtility and adds regression tests for
these issues.

Fixes: #43668
Fixes: #44312
Releases: 6.0, 6.1

Change-Id: I4a577ad749bde397a572ca48ae810e3dfe8a5740